Revert "Frontend aufräumen"
